Update: refactor api update chapter

+++ src/main/java/vn/edu/ut/service/impl/ChapterServiceImpl.java
@@ -1,80 +1,76 @@
 package vn.edu.ut.service.impl;
 
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.springframework.stereotype.Service;
 import vn.edu.ut.entity.Chapter;
 import vn.edu.ut.entity.Courses;
-import vn.edu.ut.exception.AppException;
+import vn.edu.ut.enums.ErrorCode;
+import vn.edu.ut.exception.AppApiException;
 import vn.edu.ut.exception.ResourceNotFoundException;
 import vn.edu.ut.payload.chapter.ChapterDto;
 import vn.edu.ut.repository.ChapterRepository;
 import vn.edu.ut.repository.CoursesRepository;
 import vn.edu.ut.service.ChapterService;
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
-
-import java.util.Objects;
 
 @Service
+@Transactional
+@RequiredArgsConstructor
 public class ChapterServiceImpl implements ChapterService {
-
-    @Autowired private ChapterRepository chapterRepository;
-    @Autowired private CoursesRepository coursesRepository;
-    @Autowired private ModelMapper modelMapper;
+    private final ChapterRepository chapterRepository;
+    private final CoursesRepository coursesRepository;
+    private final ModelMapper modelMapper;
 
     @Override
     public ChapterDto createChapter(Integer courseId, ChapterDto chapterDto) {
         Courses course = coursesRepository.findById(courseId)
                 .orElseThrow(() -> new ResourceNotFoundException("Course", "id", courseId));
 
-        Chapter chapterDuplicate = chapterRepository.findChapterByNameAndCourse(chapterDto.getName(), course);
+        if (chapterRepository.existsChapterByNameAndCourse(chapterDto.getName(), course))
+            throw new AppApiException(ErrorCode.DUPLICATE_NAME_CHAPTER_WITHIN_COURSE);
 
-        if(chapterDuplicate != null){
-            throw new AppException(HttpStatus.BAD_REQUEST, "Tên chương đã từng tồn tại trong khóa học " + course.getTitle());
-        }
-
-        Chapter chapter = new Chapter();
-        chapter.setName(chapterDto.getName());
-        chapter.setOrders(chapterDto.getOrders());
-        chapter.setCourse(course);
-
-        Chapter savedChapter = chapterRepository.save(chapter);
-
-        return modelMapper.map(savedChapter, ChapterDto.class);
+        Chapter chapter = new Chapter(chapterDto.getName(), chapterDto.getOrders(), course);
+        Integer maxOrderChapter = chapterRepository.findMaxOrderByCourse(courseId);
+        if (chapterDto.getOrders() <= maxOrderChapter)
+            chapterRepository.updateChapterOrderByCourseIdAfterCreateNewChapter(courseId, chapterDto.getOrders());
+        chapterRepository.save(chapter);
+        return modelMapper.map(chapter, ChapterDto.class);
     }
 
     @Override
     public ChapterDto updateChapter(Integer courseId, Integer chapterId, ChapterDto chapterDto) {
-        Courses course = coursesRepository.findById(courseId)
-                .orElseThrow(() -> new ResourceNotFoundException("Course", "id", courseId));
+        if (!coursesRepository.existsById(courseId))
+            throw new ResourceNotFoundException("Course", "id", courseId);
 
-        Chapter chapter = chapterRepository.findById(chapterId)
+        Chapter chapterInDB = chapterRepository.findById(chapterId)
                 .orElseThrow(() -> new ResourceNotFoundException("Chapter", "id", chapterId));
 
-        Chapter chapterDuplicate = chapterRepository.findChapterByNameAndCourse(chapterDto.getName(), course);
+        if (!chapterRepository.existsNameInChapterAndCourse(chapterDto.getName(), chapterId, courseId))
+            throw new AppApiException(ErrorCode.DUPLICATE_NAME_CHAPTER_WITHIN_COURSE);
 
-        if(chapterDuplicate != null){
-            if(!Objects.equals(chapter.getId(), chapterDuplicate.getId()))
-            {
-                throw new AppException(HttpStatus.BAD_REQUEST, "Tên chương này đã từng tồn tại trong khóa học " + course.getTitle());
-            }
-        }
+        chapterInDB.setName(chapterDto.getName());
 
-        chapter.setName(chapterDto.getName());
+        if (chapterDto.getOrders() > chapterInDB.getOrders())
+            chapterRepository.updateChapterOrderByCourseIdAfterUpdateChapterPushBottom(courseId,
+                    chapterInDB.getOrders(), chapterDto.getOrders());
+        else if (chapterDto.getOrders() < chapterInDB.getOrders())
+            chapterRepository.updateChapterOrderByCourseIdAfterUpdateChapterPushUp(courseId,
+                    chapterInDB.getOrders(), chapterDto.getOrders());
+        chapterInDB.setOrders(chapterDto.getOrders());
 
-        Chapter savedChapter = chapterRepository.save(chapter);
-        return modelMapper.map(savedChapter, ChapterDto.class);
+        chapterRepository.save(chapterInDB);
+        return modelMapper.map(chapterInDB, ChapterDto.class);
     }
 
     @Override
-    public String deleteChapter(Integer courseId, Integer chapterId) {
-        Courses course = coursesRepository.findById(courseId)
-                .orElseThrow(() -> new ResourceNotFoundException("Course", "id", courseId));
+    public void deleteChapter(Integer courseId, Integer chapterId) {
+        if (!coursesRepository.existsById(courseId))
+            throw new ResourceNotFoundException("Course", "id", courseId);
 
         Chapter chapter = chapterRepository.findById(chapterId)
                 .orElseThrow(() -> new ResourceNotFoundException("Chapter", "id", chapterId));
 
         chapterRepository.delete(chapter);
-        return "Xóa chương thành công!";
     }
 }
